requirements.txt
aws-cdk-lib==2.100.0
constructs>=10.0.0
boto3==1.28.85
python-dotenv==1.0.0